Fixed build error.
git-svn-id: svn://daria.forty-two.fr/shr/trunk@341 a7dca5fb-7094-41a2-a069-28cb1271ccbd

+++ openembedded/packages/images/shr-image.bb
@@ -181,7 +181,8 @@ shr_rootfs_postprocess() {
     rm ./etc/opkg/armv4t-feed.conf
     rm ./etc/opkg/fic-gta02-feed.conf
     rm ./etc/opkg/neo1973-feed.conf
-    rm ./etc/opkg/om-gta02-feed.conf
+    if [ -a ./etc/opkg/om-gta01-feed.conf ]; then rm ./etc/opkg/om-gta02-feed.conf; fi
+    if [ -a ./etc/opkg/om-gta02-feed.conf ]; then rm ./etc/opkg/om-gta02-feed.conf; fi
 
     # set openmoko theme
     mkdir -p ./etc/gtk-2.0/
